Improve C code for lock benchmarks
