+++ benchmarks/locks/mutex.c
@@ -1,0 +1,41 @@
+#include <pthread.h>
+#include <mutex.h>
+#include <assert.h>
+
+#ifndef NTHREADS
+#define NTHREADS 3
+#endif
+
+int shared;
+mutex_t mutex;
+int sum = 0;
+
+void *thread_n(void *arg)
+{
+    intptr_t index = ((intptr_t) arg);
+
+    mutex_lock(&mutex);
+    shared = index;
+    int r = shared;
+    assert(r == index);
+    sum++;
+    mutex_unlock(&mutex);
+    return NULL;
+}
+
+int main()
+{
+    pthread_t t[NTHREADS];
+
+    mutex_init(&mutex);
+
+    for (int i = 0; i < NTHREADS; i++)
+        pthread_create(&t[i], 0, thread_n, (void *)i);
+
+    for (int i = 0; i < NTHREADS; i++)
+        pthread_join(t[i], 0);
+
+    assert(sum == NTHREADS);
+
+    return 0;
+}
